Renamed assembly to BlinkStickDotNet and main class to BlinkStick
