--- BlinkStick.Hid/BlinkStick.Hid.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlinkStick.Hid", "BlinkStick.Hid.csproj", "{7AAEBEBE-E38D-47B1-A04C-A055DCEB0757}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7AAEBEBE-E38D-47B1-A04C-A055DCEB0757}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7AAEBEBE-E38D-47B1-A04C-A055DCEB0757}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7AAEBEBE-E38D-47B1-A04C-A055DCEB0757}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7AAEBEBE-E38D-47B1-A04C-A055DCEB0757}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = BlinkStick.Hid.csproj
-	EndGlobalSection
-EndGlobal
