+++ BlinkStickDotNet/AssemblyInfo.cs
