+++ BlinkStickDotNet/WinUsbDeviceMonitor.cs
@@ -21,7 +21,7 @@ using System.Runtime.InteropServices;
 using System.Security.Permissions;
 using System.Windows.Forms;
 
-namespace BlinkStick.Hid
+namespace BlinkStickDotNet
 {
 	public class WinUsbDeviceMonitor
 	{
